moved loop algo outside the class and got rid of cxx
